backend/requirements.txt
# FastAPI & server
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0

# PostGIS / spatial
geoalchemy2==0.15.2
shapely==2.0.6

# Validation & env
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Phase 6: encryption (GDPR)
cryptography==44.0.0
